feat: add api endpoint to get total count of torrents based on query keyword
The `/api/v0.1/torrentstotal` endpoint is used to get the total count of queried torrents.

When querying torrents through the `/api/v0.1/torrents` endpoint, it only returns the data for the current page without specifying how many total records there are.

The web interface, pagination is implemented by clicking "Load More". With this new endpoint, the total number of matching records can be fetched, making it easier to implement pagination in the web interface.

This will enable an implementation for pagination in the web query interface, no longer relying on 'id' for page navigation.